backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
PyJWT
flasgger
Werkzeug